Add some stubs and sbrk

+++ lib/cos/src/brk.c
@@ -1,4 +1,4 @@
 #include <arch/x86_64-cos/syscalls.h>
 #include <arch/x86_64-cos/unistd.h>
 
-int brk (void* addr) { return syscall3 (SYSCALL_SYS_BRK, (uint64_t)addr, 0, 0); }
+uint64_t brk (void* addr) { return syscall3 (SYSCALL_SYS_BRK, (uint64_t)addr, 0, 0); }
